+++ checkov/terraform/checks/utils/base_cloudsplaining_iam_scanner.py
